fix(song): ensure song isnt in an album when trying to delete song coverStorageKey

+++ backend/src/main/java/com/soundspace/service/SongCoreService.java
@@ -115,7 +115,7 @@ public class SongCoreService {
             }
 
             StorageKey coverKey = song.getCoverStorageKey();
-            if (coverKey != null && !coverKey.getId().equals(DEFAULT_COVER_IMAGE_STORAGE_KEY_ID) && coverKey.getKey() != null && !coverKey.getKey().isBlank()) {
+            if (song.getAlbum() == null && coverKey != null && !coverKey.getId().equals(DEFAULT_COVER_IMAGE_STORAGE_KEY_ID) && coverKey.getKey() != null && !coverKey.getKey().isBlank()) {
                 try {
                     storageService.delete(coverKey.getKey());
                 } catch (Exception ex) {
